requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-environ==0.11.2
python-decouple==3.8
channels==4.0.0
celery==5.3.4
redis==5.0.1
requests==2.31.0 